+++ pkgs/intl4x/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Add examples of extensions to `List` and `DateTime` formatting.
 - Removing `EcmaPolicy`, fixing Locale printing, and other small clean-ups.
+- Update to `native_toolchain_c` v0.16.7.
+- Make package compatible to run on Windows again.
 
 ## 0.12.2